# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ちょっとレンタルサービスのAPI仕様書",
        "title": "Chotto Rental API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/applications/{application_id}/matching": {
            "post": {
                "description": "応募を承認してマッチングを作成します。応募のステータスを「accepted」に、案件のステータスを「matched」に更新します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "マッチング"
                ],
                "summary": "マッチング作成",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token} 形式",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "応募ID",
                        "name": "application_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "マッチング作成情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_matching.CreateMatchingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.Matching"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "リクエストが不正です",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "403": {
                        "description": "この操作を行う権限がありません",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "404": {
                        "description": "案件が見つかりません",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    }
                }
            }
        },
...
        "/matchings/{id}/confirm-arrival": {
            "post": {
                "description": "マッチング成立後の到着確認を行います。依頼者とヘルパー両方の確認が完了すると、マッチングが完了状態になります。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "マッチング"
                ],
                "summary": "到着確認",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token} 形式",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "マッチングID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.Matching"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "リクエストが不正です / この案件は募集を終了しています / 確認期限が過ぎています",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "403": {
                        "description": "この操作を行う権限がありません",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    }
                }
            }
        },
        "/reviews": {
            "get": {
                "description": "ユーザーのレビュー一覧を取得します。user_idクエリパラメータを指定すると、特定のユーザーに対するレビューのみを取得します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "レビュー"
                ],
                "summary": "レビュー一覧取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ユーザーID（指定しない場合は全レビューを取得）",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "レビュー一覧",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.Review"
                            }
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "マッチング完了後、相手ユーザーへのレビューを作成します。1つのマッチングにつき1人のユーザーは1回のみレビュー可能です。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "レビュー"
                ],
                "summary": "レビュー作成",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token} 形式",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "レビュー情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_review.CreateReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "レビュー作成成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.Review"
                        }
                    },
                    "400": {
                        "description": "リクエストが不正です / 完了していないマッチングにはレビューできません / このマッチングに関係のないユーザーにレビューはできません",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "403": {
                        "description": "このマッチングにレビューを投稿する権限がありません",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "404": {
                        "description": "マッチングが見つかりません",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "409": {
                        "description": "既にレビューを投稿済みです",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー / 信頼スコアの更新に失敗しました",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_kageyama0_chotto-rental_internal_model.Case": {
            "description": "案件情報",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "durationMinutes": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "reward": {
                    "type": "integer"
                },
                "scheduledDate": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.User"
                },
                "userID": {
                    "description": "案件の依頼者",
                    "type": "string"
                }
            }
        },
        "github_com_kageyama0_chotto-rental_internal_model.Matching": {
            "description": "マッチング情報",
            "type": "object",
            "properties": {
                "arrivalConfirmationDeadline": {
                    "type": "string"
                },
                "arrivalConfirmedByHelper": {
                    "type": "boolean"
                },
                "arrivalConfirmedByRequester": {
                    "type": "boolean"
                },
                "case": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.Case"
                },
                "caseID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "helper": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.User"
                },
                "helperID": {
                    "description": "案件を受けたユーザー",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "meetingLocation": {
                    "type": "string"
                },
                "requester": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.User"
                },
                "requesterID": {
                    "description": "案件の依頼者",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_kageyama0_chotto-rental_internal_model.Review": {
            "description": "レビュー情報",
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "matching": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.Matching"
                },
                "matchingID": {
                    "type": "string"
                },
                "reviewedUser": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.User"
                },
                "reviewedUserID": {
                    "description": "レビューを書かれたユーザー",
                    "type": "string"
                },
                "reviewer": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.User"
                },
                "reviewerID": {
                    "description": "レビューを書いたユーザー",
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_kageyama0_chotto-rental_internal_model.User": {
            "description": "ユーザー情報",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "noShowCount": {
                    "type": "integer"
                },
                "passwordHash": {
                    "type": "string"
                },
                "trustScore": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_kageyama0_chotto-rental_pkg_util.Response": {
            "description": "HTTPレスポンス",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "internal_handler_matching.CreateMatchingRequest": {
            "type": "object",
            "required": [
                "application_id",
                "meeting_location"
            ],
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "meeting_location": {
                    "type": "string"
                }
            }
        },
        "internal_handler_review.CreateReviewRequest": {
            "type": "object",
            "required": [
                "matching_id",
                "reviewed_user_id",
                "score"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "matching_id": {
                    "type": "string"
                },
                "reviewed_user_id": {
                    "type": "string"
                },
                "score": {
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "「Bearer 」の後にJWTトークンを付与してください",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}